Initial code on Human Player (read description)
Was it formally discussed whether we will use Jupyter notebooks?

The Human implements one of many ways of giving players partial access to the state. The alternative (to assigning visible state as fields and having Game update them) would be to have in Game a getter method that identifies the player and gives appropriate visible state. The method get_state() is left as blank until more info about Game comes around.